+++ app/views/install/prerequisites.php
@@ -1,6 +1,23 @@
 <div id="install">
 
-<h2>Prerequisites</h2>
+<h1>Prerequisites</h1>
+
+<div class="help">
+
+    <?php foreach ($checks as $key => $msg): ?>
+      <div class="check-item">
+        <h2><?php echo $key ?> :</h2>
+        <?php echo $msg ?>
+      </div>
+    <?php endforeach ?>
+
+    <div class="warning">
+        <p>If you changed something to your apache or system conf, dont forget to reload apache :</p>
+        <pre>/etc/init.d/apache2 restart</pre>
+    </div>
+
+</div>
+
 
 <div class="help">
     <p>The next form will help you to configure filez by generating the file "<i>config/filez.ini</i>".</p>
@@ -10,8 +27,9 @@
 </div>
 
 <p class="submit">
-    <a href="<?php echo url_for ('/configure') ?>" class="awesome large blue">Ok, let me continue</a>
+    <a href="<?php echo url_for ('/') ?>" class="awesome large blue">Check system config again </a> or <a href="<?php echo url_for ('/configure') ?>" class="awesome large blue">continue</a>
 </p>
+
 
 </div>